+++ WorldOfTheFittest/src/grass.cpp
@@ -66,10 +66,10 @@ void Grass::draw() {
     // Set the color
     this->shader->setUniform4f("ourColor", 0.0f, 0.4f, 0.0f, 1.0f);
     // Calculate the center of the Grass to the screen range
-    double centerX = SCR_LIVING_DELTA_X * location.x - SCR_LIVING_DELTA_X / 2 - SCR_GLFW_RANGE / 2;
-    double centerY = SCR_LIVING_DELTA_Y * location.y - SCR_LIVING_DELTA_Y / 2 - SCR_GLFW_RANGE / 2;
-    double radiusX = SCR_LIVING_DELTA_X * 0.75;
-    double radiusY = SCR_LIVING_DELTA_Y * 0.75;
+    float centerX = SCR_LIVING_DELTA_X * location.x - SCR_LIVING_DELTA_X / 2 - SCR_GLFW_RANGE / 2;
+    float centerY = SCR_LIVING_DELTA_Y * location.y - SCR_LIVING_DELTA_Y / 2 - SCR_GLFW_RANGE / 2;
+    float radiusX = SCR_LIVING_DELTA_X * 0.75f;
+    float radiusY = SCR_LIVING_DELTA_Y * 0.75f;
 
     drawCircle(centerX, centerY, radiusX, radiusY);
